Move loading page via headless browser
To HeadlessBrowserLoaderHelper. A little WIP: have to rewrite a test
with complex mocking somehow.
